# Conflicts:
#	components/camel-jms/src/test/java/org/apache/camel/component/jms/ActiveMQOriginalDestinationTest.java
#	components/camel-sjms/pom.xml
#	components/camel-sjms/src/test/java/org/apache/camel/component/sjms/producer/QueueProducerQoSTest.java
#	test-infra/camel-test-infra-artemis/src/test/java/org/apache/camel/test/infra/artemis/services/AbstractArtemisEmbeddedService.java
#	test-infra/camel-test-infra-artemis/src/test/java/org/apache/camel/test/infra/artemis/services/ArtemisService.java
#	test-infra/camel-test-infra-artemis/src/test/java/org/apache/camel/test/infra/artemis/services/ArtemisServiceFactory.java

+++ components/camel-resteasy/src/main/java/org/apache/camel/component/resteasy/servlet/ResteasyServletInputStreamCopier.java
@@ -39,7 +39,7 @@ public class ResteasyServletInputStreamCopier extends ServletInputStream {
         /* create a new input stream from the cached request body */
         byte[] inputBytes = IOUtils.toByteArray(inputStream);
         this.input = new InputStreamCache(inputBytes);
-        this.copy = new CachedByteArrayOutputStream(Long.valueOf(inputBytes.length).intValue());
+        this.copy = new CachedByteArrayOutputStream(inputBytes.length);
     }
 
     @Override
